fix something dont what to fix

+++ tvpssmis/src/main/java/com/example/tvpssmis/service/ProgramService.java
@@ -55,7 +55,7 @@ public class ProgramService {
         programDAO.delete(programId);
     }
     @Transactional
-    public int getProgramCount() {
+    public long getProgramCount() {
         return programDAO.countPrograms();
     }
